requirements.txt
pandas
confluent_kafka
faker
azure-eventhub